+++ ReconNinja/core/__init__.py
@@ -1,1 +1,0 @@
-# ReconNinja v3